refactor: Separate UI Logic from domain
- 사용자 입력 관련 로직을 InputView로 분리
