+++ src/main/java/racinggame/view/InputView.java
@@ -1,36 +1,26 @@
-package racinggame.domain;
+package racinggame.view;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
+import racinggame.domain.Car;
 
-public class GameInit {
+public class InputView {
     private static final String SEPARATOR = ",";
-    private static final int MAX_LENGTH = 5;
 
-    public List<String> inputCarName() {
+    public static List<String> getCarName() {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         while (true) {
             Scanner scanner = new Scanner(System.in);
             String userInput = scanner.nextLine();
             List<String> carNames = Arrays.asList(userInput.split(SEPARATOR));
-            if (lessThanMaxLength(carNames)) {
+            if (Car.isLessThanMax(carNames)) {
                 return carNames;
             }
         }
     }
 
-    public boolean lessThanMaxLength (List<String> playerNameList) {
-        for (String playerName : playerNameList) {
-            if (playerName.length() > MAX_LENGTH) {
-                System.out.println("[ERROR] 자동차의 이름은 5자를 초과할 수 없습니다.");
-                return false;
-            }
-        }
-        return true;
-    }
-
-    public int inputTryNumber() {
+    public static int getTryNumber() {
         System.out.println("시도할 횟수는 몇 회인가요?");
         Scanner scanner = new Scanner(System.in);
         return scanner.nextInt();
